requirements.txt
pandas>=2.1.0
python-dotenv>=1.0.0

sqlalchemy>=2.0.0
snowflake-connector-python[pandas]>=3.6.0
snowflake-sqlalchemy>=1.5.0

aiohttp>=3.9.0
certifi>=2023.11.17

openai>=1.12.0
pypdf>=4.0.0
